+++ monkeytype/typing.py
@@ -3,6 +3,7 @@
 #
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
+import functools
 import inspect
 import types
 from abc import ABC, abstractmethod
@@ -499,6 +500,78 @@ class RewriteGenerator(TypeRewriter):
         if args[1] is NoneType and args[2] is NoneType:
             return Iterator[args[0]]
         return typ
+
+
+class RewriteMostSpecificCommonBase(TypeRewriter):
+    """
+    Relace a union of classes by the most specific
+    common base of its members (while avoiding multiple
+    inheritance), i.e.,
+
+    Union[Derived1, Derived2] -> Base
+    """
+
+    def _compute_bases(self, klass):
+        """
+        Return list of bases of a given class,
+        going from general (i.e., closer to object)
+        to specific (i.e., closer to class).
+        The list ends with the class itself, its
+        first element is the most general base of
+        the class up to (but excluding) any
+        base class having multiple inheritance
+        or the object class itself.
+        """
+        bases = []
+
+        curr_klass = klass
+
+        while curr_klass is not object:
+            bases.append(curr_klass)
+
+            curr_bases = curr_klass.__bases__
+
+            if len(curr_bases) != 1:
+                break
+
+            curr_klass = curr_bases[0]
+        return bases[::-1]
+
+    def _merge_common_bases(self, first_bases, second_bases):
+        """
+        Return list of bases common to both* classes,
+        going from general (i.e., closer to object)
+        to specific (i.e., closer to both classes).
+        """
+        merged_bases = []
+
+        # Only process up to shorter of the lists
+        for first_base, second_base in zip(first_bases, second_bases):
+            if first_base is second_base:
+                merged_bases.append(second_base)
+            else:
+                break
+
+        return merged_bases
+
+    def rewrite_Union(self, union):
+        """
+        Rewrite the union if possible, if no meaningful rewrite is possible,
+        return the original union.
+        """
+        klasses = union.__args__
+
+        all_bases = []
+
+        for klass in klasses:
+            klass_bases = self._compute_bases(klass)
+            all_bases.append(klass_bases)
+
+        common_bases = functools.reduce(self._merge_common_bases, all_bases)
+
+        if common_bases:
+            return common_bases[-1]
+        return union
 
 
 DEFAULT_REWRITER = ChainedRewriter(