+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 main
 ----
 
+* Add ``RewriteMostSpecificCommonBase`` type rewriter to replace union of derived classes with common
+  base class (not enabled by default.) Merge of #311, fixes #298. Thanks Christoph Hansknecht.
+
 * Drop support for Python 3.7.